refactor(plugins): split logger into lib util + plugin (CP-1)
Previously src/lib/logger/index.ts exported both the Pino logger
instance (plain utility used across services and jobs) and the
loggerPlugin (Elysia integration that adds X-Request-ID and access
log). Split into:

- src/lib/logger.ts        -> Pino logger, shouldIgnore, generateRequestId
- src/plugins/logger/
    logger-plugin.ts       -> Elysia plugin (imports from @/lib/logger)
    CLAUDE.md              -> moved from src/lib/logger/, paths updated
    __tests__/logger-plugin.test.ts (renamed from logger.test.ts)

The 17 files importing `{ logger }` from `@/lib/logger` continue to
work unchanged — the new path is exactly `@/lib/logger`. Only 4 sites
importing `{ loggerPlugin }` updated to `@/plugins/logger/logger-plugin`:
src/index.ts, src/test/support/app.ts, src/test/helpers/app.ts,
src/lib/errors/__tests__/error-plugin.test.ts.

Tests: 37 pass across plugins/logger, plugins/health, lib/errors.
