+++ src/lib/errors/__tests__/error-plugin.test.ts
@@ -3,7 +3,7 @@ import { Elysia } from "elysia";
 import { z } from "zod";
 import { AppError } from "@/lib/errors/base-error";
 import { errorPlugin, formatErrorDetail } from "@/lib/errors/error-plugin";
-import { loggerPlugin } from "@/lib/logger";
+import { loggerPlugin } from "@/plugins/logger/logger-plugin";
 
 const REQUEST_ID_PATTERN = /^req-[0-9a-f-]{36}$/;
