+++ src/plugins/logger/__tests__/logger-plugin.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from "bun:test";
 import { Elysia } from "elysia";
 import { AppError } from "@/lib/errors/base-error";
 import { errorPlugin } from "@/lib/errors/error-plugin";
-import { loggerPlugin } from "..";
+import { loggerPlugin } from "../logger-plugin";
 
 // UUIDv4: xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx (version 4)
 const UUID_V4_PATTERN =
