+++ src/test/support/app.ts
@@ -3,7 +3,6 @@ import { Elysia } from "elysia";
 import { env } from "@/env";
 import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { errorPlugin } from "@/lib/errors/error-plugin";
-import { loggerPlugin } from "@/lib/logger";
 import { adminController } from "@/modules/admin";
 import { auditController } from "@/modules/audit";
 import { employeeController } from "@/modules/employees";
@@ -12,6 +11,7 @@ import { organizationController } from "@/modules/organizations";
 import { paymentsController } from "@/modules/payments";
 import { publicController } from "@/modules/public";
 import { healthPlugin } from "@/plugins/health/health-plugin";
+import { loggerPlugin } from "@/plugins/logger/logger-plugin";
 
 /**
  * Creates a test application instance with all controllers registered.
